Migrate search methods
